[update] split files
